Add bean classes for the job-config.yaml file
the job-config.yaml is used to feed data to testgrid commands.

ATM, we are feeding loads of inputs into the ./testgrid script as input arguments. This is not scalable as the requirements grow. Hence, we need to introduce a job-config.yaml that contain this information.

This file will be the input to generate-testplan command. The input for run-testplan is a generated test-plan file. In this case, all the inputs need to be self-contained within the test-plan file.

As part of this, we are also including the infrastructureRepository etc. in the test-plan itself.