Treat InfrastructureParameter properties as first-class InfrastructureParameter instances at runtime
This is useful when two or more infrastructure parameter belong together.
Classic example is the database and database version. These two must be coupled together
and added as a single Infrastructure Parameter.

So, we will add a wrapper InfrastructureParameter called 'Database', and add the database engine
(DBEngine) and version (DBEngineVersion) as its properties.

The properties column in InfrastructureParameter table must conform the .properties file format.