Fix Dashboard Updating Issue
Since Entity Manager retrieves data from its associated cache and if we completely bypass the entity manager cache for updating the database, the state held by the cache won't be updated to the updated database state. Hence in order to avoid this inconsistency, it is required to clear the cache when we retrieve date.

Therefore, In order to get data forcefully from the database instead of caching, called refresh() method of Entity Manager for a returned entity.

Further, if List of objects returned by a query is not itself an Entity. Therefore in that kind of case, in order to refresh a bunch of things, iterate through them and refresh them individually.

This resolves wso2/testgrid#661