Add REST API for downloading reprot
It is required to have a feature to download the latest TEST GRID report for each product through the TEST GRID dashboard. Hence, in order to achieve the task, added REST API to get the latest report from the remote storage. Since TEST grid is using the AWS for remote storage, this API is able to get the latest report for a specific product from the AWS S3 bucket.