Add infra combinations filtering method for a given infra
Added new Input field. When we type or delete text in the input field it calls the handleInputChange method. Based on the text which is typed in the input field, filter the data on hits array and re-render the DOM.

If the input field is empty, it displays all infra combinations. If the input field contains text which doesn't match with any infrastructure in the table, it displays empty table.

This resolves #659