Fix the report downloading issue through the dashboard.
Since the report downloading API returns stream it doesn't convert to file when we invoke API through fetch. Hence added HEAD API and it verifies the existence of the report and it doesn't return a body as well. Thereafter check the response code of HEAD API invocation and invoke the GET API to download the report. The GET API will be invoked in the callback of the first request.

This resolves #765