Fix the alternative directory creating and scenario test running failure for the same job
In order to fix this issue modified the test plans creating command. Therefore removed 'working-dir' parameter in the test plans creating command. When creating test plans read the file path of the job-config.yaml and if it contains 'TestGrid_Home/jobs/*' pattern, create test-plans directory and test plans yaml files inside the parent directory of job-config.yaml. Otherwise, create a directory inside the TestGrid_Home/jobs and name it as <Product_Name> and use it to create and store test-plans directory and test plans.

This resolves #757