Improve TG dashboard by adding font awesome icons
TG has used images to indicate states of product builds and it affects to the performance.  Further in order to improve the overall view replaced images with font awesome icons. In addition to that infra-history page has displayed incorrect data when we view the test report page and navigate back to infra-history page and fixed it by sorting the retrieving data.

This resolves #786, resolves #787