Add last success column to the TG home page
In order to increase the sense of displaying data on the home page, removed the latest build column from the home page and add the last success column. To maintain the product status modified the finalize-test-plan command. Loop all test plan files and get status of the test plan. If one of the test plans is in the FAIL status update the last_failure_timestamp column in the product table. If all test plans are in success status update the last_success_timestamp.