Modify is_success field in the test_case table
In current TG implementation, it stores the status of the test case as boolean (i.e. True or False). But when we consider integration test there could be some skipped test cases because of failure of a dependent test case. Hence storing the status of those test cases as Failure is incorrect. Therefore modified the data type of the field and according to that code source also modified. In the new implementation maintain status column as varchar. Further, 'SUCCESS', 'FAIL' and 'SKIP' statuses are maintained.